+++ v0x04/match_fields.py
@@ -42,7 +42,10 @@ class MatchDLVLAN(MatchField):
         except ValueError:
             value, mask = map(int, self.value.split('/'))
             oxm_hasmask = True
-        value = value | VlanId.OFPVID_PRESENT
+        if value == 0:
+            value = VlanId.OFPVID_NONE
+        else:
+            value = value | VlanId.OFPVID_PRESENT
         value_bytes = value.to_bytes(2, 'big')
         if mask:
             mask = mask | VlanId.OFPVID_PRESENT
@@ -53,11 +56,20 @@ class MatchDLVLAN(MatchField):
 
     @classmethod
     def from_of_tlv(cls, tlv):
-        """Return an instance from a pyof OXM TLV."""
-        vlan_id = int.from_bytes(tlv.oxm_value[:2], 'big') & 4095
+        """Return an instance from a pyof OXM TLV.
+        *oxm_value=0x1000/0x1000 is a valid case. It's necessary to
+        deserialize 0x1000 as 4096 thus it is ignore to not obtain 0
+        """
+        vlan_id = int.from_bytes(tlv.oxm_value[:2], 'big')
+        if vlan_id != 4096:
+            # Ignore 4096
+            vlan_id &= 4095
         value = vlan_id
         if tlv.oxm_hasmask:
-            vlan_mask = int.from_bytes(tlv.oxm_value[2:], 'big') & 4095
+            vlan_mask = int.from_bytes(tlv.oxm_value[2:], 'big')
+            if vlan_mask != 4096:
+                # Ignore 4096
+                vlan_mask &= 4095
             value = f'{vlan_id}/{vlan_mask}'
         return cls(value)